most of frontier creating related functions extracted to the Frontier files
