+++ ast_explore_lite/src/FrontierSearch.cpp
@@ -12,7 +12,7 @@
 #include <explore/costmap_tools.h>
 #include <explore/FrontierSearch.h>
 #include <explore/explore_utils.h>
-
+#include "explore/Frontier.h"
 
 namespace frontier_exploration
 {
@@ -20,36 +20,6 @@ using costmap_2d::LETHAL_OBSTACLE;
 using costmap_2d::NO_INFORMATION;
 using costmap_2d::FREE_SPACE;
 
-inline double angular_vector_distance(const double x1, const double y1,const double x2,const double y2){
-    return std::abs( atan2(y1, x1) - atan2(y2, x2) ) * 180.0 / M_PI;
-}
-inline double angular_vector_distance(const geometry_msgs::Point &v1,const  geometry_msgs::Point &v2) {
-    return angular_vector_distance(v1.x, v1.y, v2.x, v2.y);
-}
-// for the case when vectors are presented by points with single point-of-origin
-inline double angular_vector_distance(const geometry_msgs::Point &p1,const  geometry_msgs::Point &p2,const  geometry_msgs::Point &reference_pt){
-    return angular_vector_distance(p1.x - reference_pt.x, p1.y - reference_pt.y, p2.x - reference_pt.x, p2.y - reference_pt.y);
-}
-
-// for now is not used
-inline geometry_msgs::Point getClosestPointTo(const std::vector<geometry_msgs::Point> &distance_vectors, const geometry_msgs::Point &goal){
-    auto min_elem_iter = std::min_element(distance_vectors.begin(), distance_vectors.end(),
-            [](const geometry_msgs::Point &p1,const geometry_msgs::Point &p2)
-            {return p1.x + p1.y < p2.x + p2.y;}); //using Manhatten distance
-    return *min_elem_iter;
-}
-
-inline geometry_msgs::Point Frontier::toReferenceFrame(const geometry_msgs::Point &pt_in_absolute_frame){
-    return makePointMsg(
-            pt_in_absolute_frame.x - reference_robot_pose.x,
-            pt_in_absolute_frame.y - reference_robot_pose.y,
-            0);
-}
-inline geometry_msgs::Point  Frontier::fromReferenceFrame(const geometry_msgs::Point &pt_in_reference_frame){
-    return makePointMsg(pt_in_reference_frame.x + reference_robot_pose.x,
-                        pt_in_reference_frame.y + reference_robot_pose.y,
-                        0);
-}
 
 FrontierSearch::FrontierSearch(costmap_2d::Costmap2D* costmap,
                                double potential_scale, double gain_scale,
@@ -64,84 +34,6 @@ FrontierSearch::FrontierSearch(costmap_2d::Costmap2D* costmap,
   , max_frontier_angular_size_(max_frontier_angular_size)
 {
 }
-    std::pair<geometry_msgs::Point, geometry_msgs::Point> FrontierSearch::approxFrontierByPlanarFarthest(Frontier &fr,
-                                                                                                         geometry_msgs::Point &reference_robot){
-        double max_dist = 0.0;
-        double heuristics_dist {0.0};
-        geometry_msgs::Point p1, p2;
-      for (auto &point: fr.vectors_to_points){
-          heuristics_dist =
-                  std::hypot(point.x , point.y ) +  /* robot_point_dist */
-                  std::hypot(point.x , point.y ); /* centroid_point dist */
-          if (heuristics_dist > max_dist){
-              max_dist = heuristics_dist;
-              p1 = point;
-          }
-      }
-      max_dist = 0.0;
-       for (auto &point: fr.vectors_to_points){
-           heuristics_dist =
-                   std::hypot(point.x - p1.x , point.y - p1.y) +  /*  point-point dist */
-                   std::hypot( (/*transforming to map frame*/point.x + reference_robot.x) - fr.centroid.x , (point.y + reference_robot.y) - fr.centroid.y); /* centroid_point dist */
-            if (heuristics_dist > max_dist){
-                max_dist = heuristics_dist;
-                p2 = point;
-            }
-        }
-        return {p1, p2};
-}
-
-
-    std::pair<geometry_msgs::Point, geometry_msgs::Point> FrontierSearch::approximateFrontierByViewAngle(
-            frontier_exploration::Frontier &fr) {
-        geometry_msgs::Point p1, p2;
-        if (!fr.vectors_to_points.empty()){
-            p1 = fr.fromReferenceFrame(fr.vectors_to_points.front());
-            p2 = fr.fromReferenceFrame(fr.vectors_to_points.back());
-            ROS_DEBUG_STREAM( " approximated borders with view angle " <<  fr.vectors_to_points.front() << fr.vectors_to_points.back());
-        } // else default zero values
-        return {p1, p2};
-    }
-
-    std::vector<Frontier> FrontierSearch::splitFrontier(Frontier& fr){
-        Frontier fr1, fr2;
-        std::vector<Frontier> frontiers{fr1, fr2}, vector_buf, res;
-        size_t splits = frontiers.size();
-        size_t current_element {0}, split_size {fr.vectors_to_points.size() / splits};
-        for (auto& this_fr: frontiers){
-            this_fr.reference_robot_pose = fr.reference_robot_pose;
-            this_fr.vectors_to_points = std::vector<geometry_msgs::Point>(
-                    fr.vectors_to_points.begin() + current_element,
-                    fr.vectors_to_points.begin() + current_element + split_size);
-
-            current_element += split_size;
-            this_fr.interpolated_line = {this_fr.vectors_to_points.front(), this_fr.vectors_to_points.back()};
-            this_fr.centroid = *(this_fr.vectors_to_points.begin() + split_size / 2);  // not true centroid, but who cares... //todo maybe
-
-            this_fr.centroid = this_fr.fromReferenceFrame(this_fr.centroid);
-            this_fr.interpolated_line.first = this_fr.fromReferenceFrame(this_fr.interpolated_line.first);
-            this_fr.interpolated_line.second = this_fr.fromReferenceFrame(this_fr.interpolated_line.second);
-
-            if (fr.vectors_to_points.end() - fr.vectors_to_points.begin() - current_element < split_size) {
-                this_fr.vectors_to_points.insert(this_fr.vectors_to_points.end(),
-                                                 fr.vectors_to_points.begin() + current_element,
-                                                 fr.vectors_to_points.end()); // uppend last chunk in case of division leftover
-            }
-
-            double fr_angular_dist = angular_vector_distance(this_fr.interpolated_line.first, this_fr.interpolated_line.second);
-            ROS_DEBUG_STREAM("FR has [" << fr_angular_dist << "]  deg  [" << frontiers[0].vectors_to_points.size() << "] PTS");
-            if (fr_angular_dist > this->max_frontier_angular_size_
-                && fr1.vectors_to_points.size() > 6)  // to allways have two subdrontiers with middlepoints
-            {
-                ROS_DEBUG_STREAM("PERFORM FRONTIER RECURSIVE SPLITTING");
-                vector_buf = splitFrontier(this_fr);
-            } else{
-                vector_buf = {this_fr};
-            }
-            res.insert(res.end(), vector_buf.begin(), vector_buf.end());
-        }
-        return res;
-    }
 
 
 std::vector<Frontier> FrontierSearch::searchFrom(geometry_msgs::Point position)
@@ -225,8 +117,8 @@ std::vector<Frontier> FrontierSearch::buildNewFrontier(unsigned int initial_cell
 {
   // initialize frontier structure
   Frontier output;
-  output.centroid.x = 0;
-  output.centroid.y = 0;
+  output.middle.x = 0;
+  output.middle.y = 0;
 
   // record initial contact point for frontier
   unsigned int ix, iy;
@@ -242,8 +134,15 @@ std::vector<Frontier> FrontierSearch::buildNewFrontier(unsigned int initial_cell
   double reference_x, reference_y;
   costmap_->indexToCells(reference, rx, ry);
   costmap_->mapToWorld(rx, ry, reference_x, reference_y);
+
   output.reference_robot_pose.x = reference_x;
   output.reference_robot_pose.y = reference_y;
+
+  FrontierParams fr_par;
+  fr_par.reference_robot_pose = makePointMsg(reference_x, reference_y);
+  fr_par.left_each = this->use_every_k_point_; // TODO pass those parameters during parameter reading
+  fr_par.max_angular_size = this->max_frontier_angular_size_;
+  fr_par.min_frontier_size = this->min_frontier_size_;
 
 size_t  cntr{0};
 
@@ -267,8 +166,9 @@ size_t  cntr{0};
             reference_scope_point.x = wx - reference_x;
             reference_scope_point.y = wy - reference_y;
             output.vectors_to_points.push_back(reference_scope_point);
-            output.centroid.x += wx; // map frame coords
-            output.centroid.y += wy;
+            fr_par.vectors_to_points.push_back(reference_scope_point);
+            output.middle.x += wx; // map frame coords
+            output.middle.y += wy;
         }
 
         // add to queue for breadth first search
@@ -277,24 +177,23 @@ size_t  cntr{0};
     }
   }
   // average out frontier centroid
-  output.centroid.x /= output.vectors_to_points.size();
-  output.centroid.y /= output.vectors_to_points.size();
+  output.middle.x /= output.vectors_to_points.size();
+  output.middle.y /= output.vectors_to_points.size();
   /*KD*/
-
   std::sort(output.vectors_to_points.begin(), output.vectors_to_points.end(),
           [](const geometry_msgs::Point &p1,const  geometry_msgs::Point &p2)
           {return atan2(p1.y, p1.x) < atan2(p2.y, p2.x);}
           );
 
   // fixme find out why there are occuring empty frontiers (empty raw points)
-  output.interpolated_line = approximateFrontierByViewAngle(output);
+  output.interpolated_line = Frontier::approximateFrontierByViewAngle(output);
   std::vector<Frontier> splitted_frontiers{output};
 if (!output.vectors_to_points.empty()){
 
     double degrees_distance = angular_vector_distance(output.interpolated_line.first, output.interpolated_line.second, output.reference_robot_pose);
     if (degrees_distance > this->max_frontier_angular_size_){
         ROS_INFO_STREAM("Detected wide frontier [" << degrees_distance <<"]" << "PTS [" <<output.vectors_to_points.size()<< "]  SPLITTING...");
-        splitted_frontiers = splitFrontier(output);
+        splitted_frontiers = Frontier::splitFrontier(output, max_frontier_angular_size_);
     }
 }
   return splitted_frontiers;
@@ -323,7 +222,7 @@ bool FrontierSearch::isNewFrontierCell(unsigned int idx,
 
 // todo Detect other room case
 bool FrontierSearch::is_hidden(frontier_exploration::Frontier &fr, double thresh_distance){
-    return std::hypot(fr.centroid.x - fr.reference_robot_pose.x, fr.centroid.y - fr.reference_robot_pose.y) < thresh_distance;
+    return std::hypot(fr.middle.x - fr.reference_robot_pose.x, fr.middle.y - fr.reference_robot_pose.y) < thresh_distance;
 }
 
 double FrontierSearch::frontierCost(const Frontier& frontier)
